requirements.txt
# Web framework
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# Auth & image handling
Pillow>=10.0
python-dotenv>=1.0

# Production server
gunicorn>=21.2

# Static files
whitenoise>=6.6

# PostgreSQL (Render)
dj-database-url>=2.1
psycopg2-binary>=2.9

# Media file storage (Cloudflare R2 / AWS S3)
django-storages[s3]>=1.14
boto3>=1.34
